Conflicts:
	src/tilda/Reverse.java
	src/tilda/db/ConnectionPool.java
	src/tilda/db/MasterFactory.java
	src/tilda/db/stores/PostgreSQL.java
	src/tilda/utils/concurrent/Executor.java
	src/tilda/utils/concurrent/SimpleRunnable.java

+++ src/tilda/utils/SystemValues.java
@@ -25,7 +25,7 @@ public abstract class SystemValues
   {
 
     /** Using -1 or 0 for an error value is so boring.... -666 is so much more attention-grabbing. */
-    public static final int           EVIL_VALUE      = -666;
+    public static final short           EVIL_VALUE      = -666;
 
     public static final ZonedDateTime STARTUP_DATE    = ZonedDateTime.now();
     public static final String        STARTUP_DATE_STR= DateTimeUtil.printDateTimeSuperCompact(STARTUP_DATE);
